Manage a separate paste buffer for middle click
When we copy, save the current clipboard prior to copying,
then when we notice the clipboard change, store that clipboard
in a separate pasteItems and restore the old clipboard.  When we
paste via middle button, set the clipboard to the stored pasteItems,
and when we can set the clipboard back to the saved current clipboard.

Because of that, restore copy-on-double-click behavior, as the typical
patterns work well with a separate clipboard and selection buffer.
There's still a little bit of a race if you select/paste too fast though?